+++ NetACS.ACS/NetACS.ACS.csproj
@@ -9,4 +9,8 @@
     <PackageReference Include="SoapCore" Version="0.9.9.8-beta" />
   </ItemGroup>
 
+  <ItemGroup>
+    <ProjectReference Include="..\NetACS.Models\NetACS.Models.csproj" />
+  </ItemGroup>
+
 </Project>
