+++ NetACS.Models/EventStruct.cs
@@ -1,0 +1,11 @@
+﻿using System.Runtime.Serialization;
+
+namespace NetACS.Models
+{
+    [DataContract]
+    public class EventStruct
+    {
+        [DataMember]
+        public string EventCode { get; set; }
+    }
+}
