+++ NetACS.sln
@@ -16,6 +16,8 @@ Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "NetACS.ACS", "NetACS.ACS\Ne
 EndProject
 Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "NetACS.Database", "NetACS.Database\NetACS.Database.csproj", "{6AFD0A0A-15B7-4C59-8549-D4061773FD3F}"
 EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "NetACS.Models", "NetACS.Models\NetACS.Models.csproj", "{5CCE429F-B58E-4F16-B49D-59D49EF0B828}"
+EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
 		Debug|Any CPU = Debug|Any CPU
@@ -34,6 +36,10 @@ Global
 		{6AFD0A0A-15B7-4C59-8549-D4061773FD3F}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{6AFD0A0A-15B7-4C59-8549-D4061773FD3F}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{6AFD0A0A-15B7-4C59-8549-D4061773FD3F}.Release|Any CPU.Build.0 = Release|Any CPU
+		{5CCE429F-B58E-4F16-B49D-59D49EF0B828}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{5CCE429F-B58E-4F16-B49D-59D49EF0B828}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{5CCE429F-B58E-4F16-B49D-59D49EF0B828}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{5CCE429F-B58E-4F16-B49D-59D49EF0B828}.Release|Any CPU.Build.0 = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
